requirements.txt
absl-py==1.3.0
addict==2.4.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
future==0.18.2
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
idna==3.4
importlib-metadata==6.0.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mmcls==0.25.0
mmcv-full==1.7.1
-e git+https://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@github.com/sanath-tech/FuseFormer.git@aa8dbc2389a8bd63fdbe7e1725e9e5d14b487837#egg=mmsegmentation
model-index==0.1.11
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.7.0.68
openmim==0.3.4
ordered-set==4.1.0
packaging==22.0
pandas==1.5.2
Pillow==9.4.0
prettytable==3.5.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.0.0
rsa==4.9
six==1.16.0
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2022.10.10
torch==1.13.1
torchvision==0.14.1
typing-extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
yapf==0.32.0
zipp==3.11.0
